Fix the Access control in the WMS groups
- Reactivate the tests
- Split the tests in more then one file
- Add WMS groups tests
- Use standard test font
