+++ .ci/travis/linux/blacklist.txt
@@ -22,12 +22,8 @@ qgis_ziplayertest
 # Flaky, see https://dash.orfeo-toolbox.org/testDetails.php?test=63061783&build=297405
 PyQgsSpatialiteProvider
 
-# Flaky, see https://travis-ci.org/qgis/QGIS/jobs/297708174
-PyQgsServerAccessControl
-
 # Need a local postgres installation
 PyQgsAuthManagerPKIPostgresTest
 PyQgsAuthManagerPasswordPostgresTest
 PyQgsAuthManagerOgrPostgresTest
 PyQgsDbManagerPostgis
-
