+++ tests/src/python/CMakeLists.txt
@@ -263,7 +263,11 @@ IF (WITH_SERVER)
   ADD_PYTHON_TEST(PyQgsServerSettings test_qgsserver_settings.py)
   ADD_PYTHON_TEST(PyQgsServerProjectUtils test_qgsserver_projectutils.py)
   ADD_PYTHON_TEST(PyQgsServerSecurity test_qgsserver_security.py)
-  ADD_PYTHON_TEST(PyQgsServerAccessControl test_qgsserver_accesscontrol.py)
+  ADD_PYTHON_TEST(PyQgsServerAccessControlWMS test_qgsserver_accesscontrol_wms.py)
+  ADD_PYTHON_TEST(PyQgsServerAccessControlWMSGetlegendgraphic test_qgsserver_accesscontrol_wms_getlegendgraphic.py)
+  ADD_PYTHON_TEST(PyQgsServerAccessControlWFS test_qgsserver_accesscontrol_wfs.py)
+  ADD_PYTHON_TEST(PyQgsServerAccessControlWCS test_qgsserver_accesscontrol_wcs.py)
+  ADD_PYTHON_TEST(PyQgsServerAccessControlWFSTransactional test_qgsserver_accesscontrol_wfs_transactional.py)
   ADD_PYTHON_TEST(PyQgsServerWFS test_qgsserver_wfs.py)
   ADD_PYTHON_TEST(PyQgsServerWFST test_qgsserver_wfst.py)
   ADD_PYTHON_TEST(PyQgsOfflineEditingWFS test_offline_editing_wfs.py)
